Refactor CIDR logic
