+++ newsfragments/1420.bugfix
@@ -1,0 +1,2 @@
+Avoid generating global networks when guessing podsCIDR and servicesCIDR for container and VPN sshuttle routing configuration.
+
